// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/sales/models/contract.model.ts
export class ContractCostDetail {
  ContractCostDetailId: string;
  CostId: string;
  ContractId: string;
  Quantity: number;
  UnitPrice: number;
  CostName: string;
  CostCode: string;
  Active: boolean;
  CreatedById: string;
  CreatedDate: Date;
  UpdatedById: string;
  UpdatedDate: Date;
  SumAmount: number;
  IsInclude: boolean;

  constructor() {
    this.ContractCostDetailId = '00000000-0000-0000-0000-000000000000',
    this.CostId = '',
    this.ContractId = '00000000-0000-0000-0000-000000000000',
    this.Quantity = 0,
    this.UnitPrice = 0,
    this.CostName = '',
    this.Active = true,
    this.CreatedById = '00000000-0000-0000-0000-000000000000',
    this.CreatedDate = new Date(),
    this.UpdatedById = null,
    this.UpdatedDate = null,
    this.IsInclude = true
  }
}